cueparser/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# cueparser/cueparser.py: 34
pytaglib == 1.4.1
